% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generar.factores.inflacion}
\alias{generar.factores.inflacion}
\title{A Deflactor Function}
\usage{
generar.factores.inflacion(archivo.csv, fecha.valuacion, vector.fechas,
  vector.montos)
}
\arguments{
\item{archivo.csv,}{fecha.valuacion, vector.fechas, vector.montos}
}
\description{
This function allows you to calculate the deflactor index.
}
\examples{
deflactacion_function()
}
\keyword{Deflact}

